Moved functions to logic layer. Removed some code small spelling errors
