+++ Logic_layer/EmployeeLogic.py
@@ -25,10 +25,10 @@ class EmployeeLogic:
     def change_employee_info(self, ssn: str, field: str, new_value: str) -> str:
         self.data_wrapper.update_employee(ssn, field, new_value)
         return "Employee information updated successfully."
-
-    def search_employee_by_id(self, ssn: str) -> Employee:
+    
+    def search_employee_by_ssn(self, ssn):
         employees = self.data_wrapper.list_employees()
         for employee in employees:
             if employee.ssn == ssn:
                 return employee
-        return None
+        raise ValueError(f"Employee with SSN {ssn} not found.")
